+++ cli/test/clojure_lsp/server_test.clj
@@ -1,7 +1,7 @@
 (ns clojure-lsp.server-test
   (:require
    [clojure-lsp.server :as server]
-   [clojure-lsp.test-helper :as h]
+   [clojure-lsp.test-helper.internal :as h]
    [clojure.test :refer [deftest is testing]]))
 
 (h/reset-components-before-test)