+++ lib/test/clojure_lsp/feature/add_missing_libspec_test.clj
@@ -3,7 +3,7 @@
    [babashka.fs :as fs]
    [clojure-lsp.feature.add-missing-libspec :as f.add-missing-libspec]
    [clojure-lsp.shared :as shared]
-   [clojure-lsp.test-helper :as h]
+   [clojure-lsp.test-helper.internal :as h]
    [clojure.java.io :as io]
    [clojure.test :refer [deftest is testing]]
    [rewrite-clj.zip :as z]))