+++ lib/test/clojure_lsp/feature/code_lens_test.clj
@@ -2,7 +2,7 @@
   (:require
    [clojure-lsp.feature.code-lens :as f.code-lens]
    [clojure-lsp.shared :as shared]
-   [clojure-lsp.test-helper :as h]
+   [clojure-lsp.test-helper.internal :as h]
    [clojure.test :refer [deftest is testing]]))
 
 (h/reset-components-before-test)