+++ lib/test/clojure_lsp/feature/diagnostics_test.clj
@@ -4,7 +4,7 @@
    [clojure-lsp.feature.diagnostics :as f.diagnostic]
    [clojure-lsp.feature.diagnostics.custom :as f.diagnostics.custom]
    [clojure-lsp.shared :as shared]
-   [clojure-lsp.test-helper :as h]
+   [clojure-lsp.test-helper.internal :as h]
    [clojure.core.async :as async]
    [clojure.test :refer [deftest is testing]]))