+++ test-helper/build.clj
@@ -4,7 +4,7 @@
    [clojure.string :as string]
    [clojure.tools.build.api :as b]))
 
-(def lib 'com.github.clojure-lsp/clojure-lsp)
+(def lib 'com.github.clojure-lsp/clojure-lsp-test-helper)
 (def current-version (string/trim (slurp (io/resource "CLOJURE_LSP_VERSION"))))
 (def class-dir "target/classes")
 (def basis {:project "deps.edn"})
@@ -13,22 +13,21 @@
 (defn clean [_]
   (b/delete {:path "target"}))
 
-(defn pom [opts]
+(defn pom [_opts]
   (b/write-pom {:class-dir class-dir
                 :lib lib
                 :src-pom "./pom.xml"
                 :version current-version
-                :basis (b/create-basis (update basis :aliases concat (:extra-aliases opts)))
+                :basis (b/create-basis basis)
                 :src-dirs ["src"]
-                :resource-dirs ["resources"]
                 :scm {:tag current-version}})
   (b/copy-file {:src (str class-dir "/META-INF/maven/" lib "/pom.xml") :target "pom.xml"})
   (b/copy-file {:src (str class-dir "/META-INF/maven/" lib "/pom.properties") :target "pom.properties"}))
 
 (defn jar [opts]
-  (clean nil)
+  (clean opts)
   (pom opts)
-  (b/copy-dir {:src-dirs ["src" "resources"]
+  (b/copy-dir {:src-dirs ["src"]
                :target-dir class-dir})
   (b/jar {:class-dir class-dir
           :jar-file file}))