+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Considerably improve performance of `unused-public-var` and `different-aliases` linters.
   - Bump rewrite-clj to 1.2.50.
   - New feature: Add support for custom project linters. #2043
+  - Publish to clojars `com.github.clojure-lsp/clojure-lsp-test-helper` to be able to test created custom linters.
 
 - Editor
   - Add support for LSP feature `textDocument/selectionRange`. #1961